refactor: API 폴더 → types 폴더 변경
