--- src/api/Item.type.ts
@@ -1,10 +1,0 @@
-export type ItemType = {
-  idx: number;
-  name: string;
-  mainImage: string;
-  description: string;
-  spaceCategory: string;
-  price: number;
-  maximumPurchases: number;
-  registrationDate: string;
-};
